fix: user 엔티티의 id 속성 데이터 타입 Long에서 Integer로 수정

+++ src/main/java/com/example/pohab/Login/Filter/JwtRequestFilter.java
@@ -46,13 +46,13 @@ public class JwtRequestFilter extends OncePerRequestFilter {
 
         String token = jwtHeader.replace(JwtProperties.TOKEN_PREFIX, "");
 
-        Long userId = null;
+        Integer userId = null;
         String userEmail = null;
         String userName = null;
 
         try {
             userId = JWT.require(Algorithm.HMAC512(JwtProperties.SECRET)).build().verify(token)
-                    .getClaim("id").asLong();
+                    .getClaim("id").asInt();
 
             userEmail = JWT.require(Algorithm.HMAC512(JwtProperties.SECRET)).build().verify(token)
                     .getClaim("email").asString();
